requirements/requirements_dev.txt
-r requirements.txt
psutil==5.9.7
coverage==6.4.2
Django==4.0.6
requests==2.28.1
djangorestframework==3.13.1